static-container-registry: server `Docker-Content-Digest` header
It appears `containerd` relies on this header to be present (though
that's not in line with the Docker Distribution spec, which declares
this header to be optional).

This requires us to generate a `location` per manifest, or I guess this
could be implemented using some map/lookup table as well.

Furthermore, `containerd` uses this header from a `HEAD` request to then
retrieve the actual manifest file *by digest*, which is now implemented
as well. As a useful side-effect, we now support by-digest pulls.

See: https://github.com/containerd/containerd/issues/3238
See: https://github.com/containerd/containerd/issues/2994